Add files via upload
This project is a replica of the Myntra e-commerce website, developed using HTML, CSS, and JavaScript. It includes essential features like product listings, search functionality, and a responsive design to ensure a seamless user experience across devices. The site simulates the look and feel of a professional online shopping platform, showcasing products with images, descriptions, and prices, offering an interactive and engaging user interface.